msm: clock-krait-8974: Expose PVS table data through module parameters
For debugging purposes, it's sometimes useful to retrieve part
PVS and speed bin data at runtime. Export this data via a module
parameters for easy access. Having a static copy of this data also
makes its recovery from a ramdump convenient.

A 'qcom,pvs-config-ver' device-tree property is also introduced.
It will not be used by software, but can be queried to see the
which version of the voltage configuration data is present in
the build.

Change-Id: I233919d8c4970fe4e1263423ab9f08cabcc0986d
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>